les affectés + exception
