+++ src/main/java/com/camp/havenfort_dev/exception/UserNotFoundException.java
@@ -1,0 +1,7 @@
+package com.camp.havenfort_dev.exception;
+
+public class UserNotFoundException extends RuntimeException{
+    public UserNotFoundException(String message) {
+        super(message);
+    }
+}
