+++ src/main/java/com/camp/havenfort_dev/repositories/ActivityRepo.java
@@ -1,6 +1,6 @@
-package repositories;
+package com.camp.havenfort_dev.repositories;
 
-import entities.Activity;
+import com.camp.havenfort_dev.entities.Activity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ActivityRepo extends JpaRepository<Activity,Long> {
