+++ src/main/java/com/camp/havenfort_dev/repositories/CampsiteRepo.java
@@ -1,6 +1,6 @@
-package repositories;
+package com.camp.havenfort_dev.repositories;
 
-import entities.Campsite;
+import com.camp.havenfort_dev.entities.Campsite;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CampsiteRepo extends JpaRepository<Campsite,Long> {
