+++ src/main/java/com/camp/havenfort_dev/repositories/CenterOfCampRepo.java
@@ -1,6 +1,6 @@
-package repositories;
+package com.camp.havenfort_dev.repositories;
 
-import entities.CenterOfCamp;
+import com.camp.havenfort_dev.entities.CenterOfCamp;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CenterOfCampRepo extends JpaRepository<CenterOfCamp,Long> {
