+++ src/main/java/com/camp/havenfort_dev/repositories/EventRepo.java
@@ -1,6 +1,6 @@
-package repositories;
+package com.camp.havenfort_dev.repositories;
 
-import entities.Event;
+import com.camp.havenfort_dev.entities.Event;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface EventRepo extends JpaRepository<Event,Long> {
