+++ src/main/java/com/camp/havenfort_dev/repositories/ReservationRepo.java
@@ -1,6 +1,6 @@
-package repositories;
+package com.camp.havenfort_dev.repositories;
 
-import entities.Reservation;
+import com.camp.havenfort_dev.entities.Reservation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ReservationRepo extends JpaRepository<Reservation,Long> {
