+++ src/main/java/com/camp/havenfort_dev/services/ICenterOfCampService.java
@@ -1,6 +1,7 @@
-package services;
+package com.camp.havenfort_dev.services;
 
-import entities.CenterOfCamp;
+import com.camp.havenfort_dev.entities.Campsite;
+import com.camp.havenfort_dev.entities.CenterOfCamp;
 
 import java.io.Serializable;
 import java.util.List;
@@ -12,4 +13,5 @@ public interface ICenterOfCampService extends Serializable {
     CenterOfCamp updateCenterOfCamp(CenterOfCamp centerOfCamp);
     CenterOfCamp findById(Long id);
     void deleteCenterOfCamp(Long id);
+    void addCenterOfCampToEvent(CenterOfCamp centerOfCamp, Long idEvent);
 }
