+++ src/main/java/com/camp/havenfort_dev/services/IEventService.java
@@ -1,6 +1,7 @@
-package services;
+package com.camp.havenfort_dev.services;
 
-import entities.Event;
+import com.camp.havenfort_dev.entities.Event;
+import com.camp.havenfort_dev.entities.Reservation;
 
 import java.io.Serializable;
 import java.util.List;
@@ -13,4 +14,7 @@ public interface IEventService extends Serializable {
     public Event updateEvent(Event event);
     public Event findById(Long id);
     public void deleteEvent(Long id);
+    void addEventToActivity(Event event,Long idActivity);
+    void addEventToReservation(Event event, Long idReservation);
+
 }
