+++ src/main/java/com/camp/havenfort_dev/services/IReservationService.java
@@ -1,6 +1,6 @@
-package services;
+package com.camp.havenfort_dev.services;
 
-import entities.Reservation;
+import com.camp.havenfort_dev.entities.Reservation;
 
 import java.io.Serializable;
 import java.util.List;
