+++ src/main/java/com/camp/havenfort_dev/services/ReservationService.java
@@ -1,9 +1,10 @@
-package services;
+package com.camp.havenfort_dev.services;
 
-import entities.Reservation;
+import com.camp.havenfort_dev.exception.UserNotFoundException;
+import com.camp.havenfort_dev.repositories.ReservationRepo;
+import com.camp.havenfort_dev.entities.Reservation;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import repositories.ReservationRepo;
 
 import java.util.List;
 @Service
@@ -27,7 +28,8 @@ public class ReservationService implements IReservationService{
 
     @Override
     public Reservation findReservationById(Long id) {
-        return null;
+        return reservationRepo.findById(id)
+                .orElseThrow(() -> new UserNotFoundException("User by id " + id + " was not found"));
     }
 
     @Override
