--- src/main/java/services/EventService.java
@@ -1,38 +1,0 @@
-package services;
-
-import entities.Event;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-import repositories.EventRepo;
-
-import java.util.List;
-@Service
-public class EventService implements IEventService{
-    @Autowired
-    EventRepo eventRepo;
-    @Override
-    public Event addEvent(Event event) {
-        return eventRepo.save(event);
-    }
-
-    @Override
-    public List<Event> findAllEvent() {
-        return eventRepo.findAll();
-    }
-
-    @Override
-    public Event updateEvent(Event event) {
-        return eventRepo.save(event);
-    }
-
-    @Override
-    public Event findById(Long id) {
-        return null;
-    }
-
-    @Override
-    public void deleteEvent(Long id) {
-        eventRepo.deleteById(id);
-
-    }
-}
