+++ src/main/resources/application.properties
@@ -1,7 +1,9 @@
 
+
+
 #Server configuration
 server.servlet.context-path=/havenfort
-server.port=8083
+server.port=8089
 
 ### DATABASE ###
 #spring.datasource.driver-class-name=com.mysql.jdbc.Driver
